+++ doc/version_history.rst
@@ -6,6 +6,16 @@
 Version History
 ###############
 
+v1.3.0
+-------
+
+* Monitor the camera cable wrap is following or not.
+
+v1.2.11
+-------
+
+* Adapt the ruff format.
+
 v1.2.10
 -------